Move mergeProperty into Visualizer
